drivers: npmx: shell: move ship code to the separate file
