+++ drivers/npmx/CMakeLists.txt
@@ -55,6 +55,7 @@ if(CONFIG_NPMX_SHELL)
     zephyr_library_sources(shell/ldsw.c)
     zephyr_library_sources(shell/led.c)
     zephyr_library_sources(shell/pof.c)
+    zephyr_library_sources(shell/ship.c)
 endif()
 
 # Get npmx include file
